refactor to extract common functions and classes out
